[FIX]: Recommend 수정

+++ src/main/java/com/example/ootd/domain/clothes/dto/data/ScoredClothesDto.java
@@ -20,6 +20,6 @@ public record ScoredClothesDto(
     Integer temperatureSensitivity,
     String thickness,
     String color,
-    Integer score
+    Double score
 ) {
 }
